code_analysis_runner.py
import clr  # Provided by pythonnet

# Load Roslyn assemblies (paths might need to be adjusted based on your setup)
clr.AddReference('Microsoft.CodeAnalysis')
clr.AddReference('Microsoft.CodeAnalysis.CSharp')

from Microsoft.CodeAnalysis import CSharp

# Example: Analyzing a simple piece of C# code
csharp_code = """
public class Program
{
    public static void Main(string[] args)
    {
        System.Console.WriteLine("Hello, World!");
    }
}
"""

# Parse the C# code into a syntax tree
syntax_tree = CSharp.CSharpSyntaxTree.ParseText(csharp_code)

# Here you would proceed to create a compilation, run analyzers, etc.