requirements.txt
GitPython
flake8
openai
